Fix scalars serialization
`graphql-java` turns the input scalars into our own wrapper types on the
way in. Sadly, though, our `GraphQLInputMapper` configuration for Jackson
doesn't know how to treat those scalars when constructing objects that
contain them. It ends up breaking down our scalars into a hash, so we
have to handle that incoming property hash to let it deserialize
properly.

This could also be handled via custom deserializers with the object
mapper, but I prefer the approach where the objects themselves have that
responsibility to avoid having to do lots of customization on the
`ObjectMapper` that out `GraphQLInputMapper` uses.